Systemd Override for Unbound
Unbound won't start if one of it's listening addresses/interfaces
is unavaible when a start is attempted. The default limits on
retrying/restarting are too restrictive in our environment where
it apparently takes more time for the ethernet interface to come
up after the host reboot. This override inreases the restart time
to 5 seconds and gives it more opportunities to restart the daemon.